Add PayPal and Phan trang

+++ ViewModel/ViewModels/OrdersVm.cs
@@ -16,7 +16,7 @@ namespace ViewModel.ViewModels
 
         public DateTime OrderDay { get; set; }
 
-        public int TotalPice { get; set; }
+        public decimal TotalPice { get; set; }
 
         public Status Status { get; set; }
